efficientdet_aituring/docker_requirements.txt
lxml==4.6.3
Pillow==8.3.2
pycocotools==2.0.1; sys_platform=='linux'
pycocotools-windows==2.0.0.2; sys_platform=='win32'
opencv-python==4.5.3.56
PyYAML==5.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboard-plugin-profile==2.5.0
tensorflow-addons==0.14.0
tensorflow-hub==0.12.0
tensorflow-estimator==2.6.0
tensorflow-model-optimization==0.5.1.dev0
matplotlib==3.4.3
neural-structured-learning==1.3.1
scikit-learn==1.0
wget==3.2; sys_platform=='win32'